create

+++ 0x15-file_io/1-create_file.c
@@ -14,7 +14,7 @@ int create_file(const char *filename, char *text_content)
 
 {
 	int fd, i;
-	int writer;
+	ssize_t writer;
 
 	if (filename == NULL)
 	return (-1);
